feat(kvevents): add dedup observability and tighten tests
Address review feedback on the KV-event dedup filter:
- Add kvcache_kvevents_dedup_removed_hashes_{suppressed,forwarded}_total
  counters (block-hash granularity, not event count) so suppressed removals are
  observable, following the pkg/kvcache/metrics global-collector pattern; emit a
  TRACE log when removals are suppressed.
- Document why the filter lives in the Pool rather than as an Index decorator.
- Tighten the concurrency doc: the mutex guards the cross-pod map and same-pod
  events are serialized by Pool.AddTask sharding.
- Move the pool-level tests into pool_test.go and add a negative test that a
  device-tier update resolving no keys is not reference-counted.
