+++ pkg/kvevents/event_dedup_filter_test.go
@@ -1,14 +1,9 @@
-package kvevents //nolint:testpackage // tests use unexported eventDedupFilter and processEventBatch
+package kvevents //nolint:testpackage // tests use the unexported eventDedupFilter
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/llm-d/llm-d-kv-cache/pkg/kvcache/kvblock"
-	"github.com/llm-d/llm-d-kv-cache/pkg/utils/logging"
 )
 
 func testScope(pod, tier string, group, dpRank int) blockScope {
@@ -191,187 +186,4 @@ func TestGroupIdxOrNoGroup(t *testing.T) {
 	assert.Equal(t, noGroupIdx, groupIdxOrNoGroup(nil))
 	g := 3
 	assert.Equal(t, 3, groupIdxOrNoGroup(&g))
-}
-
-// TestPool_DuplicateStoreSurvivesFirstRemove is the end-to-end proof through
-// processEventBatch with a real index: two overlapping chunks announce the same
-// blocks, so the first BlockRemoved must not evict them and the second must.
-func TestPool_DuplicateStoreSurvivesFirstRemove(t *testing.T) {
-	ctx := logging.NewTestLoggerIntoContext(context.Background())
-	pool, idx, tp := newTestPool(t, 16)
-
-	tokens := makeTokens(64)
-	engineKeys := makeEngineKeys(4, 800)
-
-	store := func() {
-		pool.processEventBatch(ctx, &EventBatch{
-			Events: []GenericEvent{
-				&BlockStoredEvent{BlockHashes: engineKeys, Tokens: tokens, ParentHash: 0},
-			},
-		}, "pod-dup", "test-model")
-	}
-	remove := func() {
-		pool.processEventBatch(ctx, &EventBatch{
-			Events: []GenericEvent{
-				&BlockRemovedEvent{BlockHashes: engineKeys},
-			},
-		}, "pod-dup", "test-model")
-	}
-
-	store() // overlapping chunk A announces these constituent hashes
-	store() // overlapping chunk B re-announces the same hashes
-
-	canonicalKeys, err := tp.TokensToKVBlockKeys(kvblock.EmptyBlockHash, tokens, "test-model", nil)
-	require.NoError(t, err)
-	require.Len(t, canonicalKeys, 4)
-
-	// First remove (chunk A evicted) must NOT drop the blocks.
-	remove()
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		require.Len(t, result[ck], 1, "block must survive the first of two duplicate removes")
-	}
-
-	// Second remove (chunk B evicted) drops them.
-	remove()
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		assert.Empty(t, result[ck], "block must be evicted after the second duplicate remove")
-	}
-
-	// Engine->request mapping should also be gone after full eviction.
-	_, err = idx.GetRequestKey(ctx, kvblock.BlockHash(engineKeys[0]))
-	assert.Error(t, err, "engine->request mapping should be removed after the second remove")
-}
-
-// TestPool_DuplicateCPUOffloadRemovalSurvivesFirstRemove exercises the
-// offloading wiring end to end through processEventBatch: a CPU tier is
-// established via the empty-token device-tier update path, re-announced by two
-// overlapping chunks, and must require two CPU removes before the CPU entry is
-// evicted — while the independently reference-counted GPU entry (never removed)
-// is untouched throughout. This jointly exercises handleDeviceTierUpdate, the
-// deviceTier normalization on both store and remove, and the dedup scope.
-func TestPool_DuplicateCPUOffloadRemovalSurvivesFirstRemove(t *testing.T) {
-	ctx := logging.NewTestLoggerIntoContext(context.Background())
-	pool, idx, tp := newTestPool(t, 16)
-
-	tokens := makeTokens(64)
-	engineKeys := makeEngineKeys(4, 900)
-
-	// Step 1: GPU store with tokens establishes the engine->request mapping
-	// that the empty-token CPU offload path resolves against.
-	pool.processEventBatch(ctx, &EventBatch{
-		Events: []GenericEvent{
-			&BlockStoredEvent{BlockHashes: engineKeys, Tokens: tokens, ParentHash: 0},
-		},
-	}, "pod-offload", "test-model")
-
-	// Step 2: the same CPU offload (empty tokens, "CPU" tier) announced twice,
-	// as two overlapping chunks would re-announce the shared constituent hashes.
-	cpuStore := func() {
-		pool.processEventBatch(ctx, &EventBatch{
-			Events: []GenericEvent{
-				&BlockStoredEvent{BlockHashes: engineKeys, Tokens: nil, ParentHash: 0, DeviceTier: "CPU"},
-			},
-		}, "pod-offload", "test-model")
-	}
-	cpuStore()
-	cpuStore()
-
-	canonicalKeys, err := tp.TokensToKVBlockKeys(kvblock.EmptyBlockHash, tokens, "test-model", nil)
-	require.NoError(t, err)
-	require.Len(t, canonicalKeys, 4)
-
-	// Both gpu and cpu entries should now exist for each canonical key.
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		require.Len(t, result[ck], 2, "gpu and cpu entries should both exist after offload")
-	}
-
-	cpuRemove := func() {
-		pool.processEventBatch(ctx, &EventBatch{
-			Events: []GenericEvent{
-				&BlockRemovedEvent{BlockHashes: engineKeys, DeviceTier: "CPU"},
-			},
-		}, "pod-offload", "test-model")
-	}
-
-	// Step 3: the first CPU remove (one overlapping chunk evicted) must be
-	// suppressed — the cpu entry still has an outstanding reference.
-	cpuRemove()
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		require.Len(t, result[ck], 2, "cpu entry must survive the first of two duplicate CPU removes")
-		tiers := map[string]bool{}
-		for _, pe := range result[ck] {
-			tiers[pe.DeviceTier] = true
-		}
-		assert.True(t, tiers["cpu"], "cpu entry should still be present after the first CPU remove")
-		assert.True(t, tiers["gpu"], "gpu entry should be untouched")
-	}
-
-	// Step 4: the second CPU remove releases the last reference and evicts the
-	// cpu entry; the gpu entry (never removed) must remain.
-	cpuRemove()
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		require.Len(t, result[ck], 1, "only the gpu entry should remain after the second CPU remove")
-		assert.Equal(t, "gpu", result[ck][0].DeviceTier, "surviving entry must be the untouched gpu copy")
-	}
-}
-
-// TestPool_AllBlocksClearedResetsDedup verifies the filter is reset on
-// AllBlocksCleared, so a post-clear store/remove cycle behaves freshly rather
-// than carrying a stale reference that would suppress the remove.
-func TestPool_AllBlocksClearedResetsDedup(t *testing.T) {
-	ctx := logging.NewTestLoggerIntoContext(context.Background())
-	pool, idx, tp := newTestPool(t, 16)
-
-	tokens := makeTokens(64)
-	engineKeys := makeEngineKeys(4, 850)
-
-	storeTwice := func() {
-		for range 2 {
-			pool.processEventBatch(ctx, &EventBatch{
-				Events: []GenericEvent{
-					&BlockStoredEvent{BlockHashes: engineKeys, Tokens: tokens, ParentHash: 0},
-				},
-			}, "pod-clr", "test-model")
-		}
-	}
-
-	// Two stores -> reference count 2 -> would normally need two removes.
-	storeTwice()
-
-	// Clear wipes both the index and the dedup counts for the pod.
-	pool.processEventBatch(ctx, &EventBatch{
-		Events: []GenericEvent{&AllBlocksClearedEvent{}},
-	}, "pod-clr", "test-model")
-
-	// Re-establish a single reference after the clear.
-	pool.processEventBatch(ctx, &EventBatch{
-		Events: []GenericEvent{
-			&BlockStoredEvent{BlockHashes: engineKeys, Tokens: tokens, ParentHash: 0},
-		},
-	}, "pod-clr", "test-model")
-
-	canonicalKeys, err := tp.TokensToKVBlockKeys(kvblock.EmptyBlockHash, tokens, "test-model", nil)
-	require.NoError(t, err)
-
-	// A single remove must now fully evict: if the pre-clear count of 2 had
-	// survived, this remove would be suppressed and the blocks would linger.
-	pool.processEventBatch(ctx, &EventBatch{
-		Events: []GenericEvent{&BlockRemovedEvent{BlockHashes: engineKeys}},
-	}, "pod-clr", "test-model")
-
-	for _, ck := range canonicalKeys {
-		result, err := idx.Lookup(ctx, []kvblock.BlockHash{ck}, nil)
-		require.NoError(t, err)
-		assert.Empty(t, result[ck], "single remove after clear must evict (dedup count was reset)")
-	}
 }
